Move worked example label providers out of ApplicationModule
