+++ app/BUILD.bazel
@@ -691,6 +691,7 @@ kt_android_library(
         "//third_party:org_jetbrains_kotlin_kotlin-stdlib-jdk8_jar",
         "//utility",
         "//utility/src/main/java/org/oppia/android/util/extensions:bundle_extensions",
+        "//utility/src/main/java/org/oppia/android/util/parser/html:worked_example_label_string_ids",
         "//utility/src/main/java/org/oppia/android/util/parser/image:image_loader",
         "//utility/src/main/java/org/oppia/android/util/parser/image:image_parsing_annonations",
         "//utility/src/main/java/org/oppia/android/util/profile:current_user_profile_id_intent_decorator",
