+++ scripts/assets/test_file_exemptions.textproto
@@ -1069,6 +1069,10 @@ test_file_exemption {
   test_file_not_required: true
 }
 test_file_exemption {
+  exempted_file_path: "app/src/main/java/org/oppia/android/app/html/WorkedExampleLabelsModule.kt"
+  test_file_not_required: true
+}
+test_file_exemption {
   exempted_file_path: "app/src/main/java/org/oppia/android/app/home/HomeActivity.kt"
   source_file_is_incompatible_with_code_coverage: true
 }
@@ -4741,6 +4745,10 @@ test_file_exemption {
 }
 test_file_exemption {
   exempted_file_path: "utility/src/main/java/org/oppia/android/util/parser/html/PolicyType.kt"
+  test_file_not_required: true
+}
+test_file_exemption {
+  exempted_file_path: "utility/src/main/java/org/oppia/android/util/parser/html/WorkedExampleQuestionLabelStringId.kt"
   test_file_not_required: true
 }
 test_file_exemption {
