+++ utility/src/main/java/org/oppia/android/util/parser/html/WorkedExampleQuestionLabelStringId.kt
@@ -1,0 +1,11 @@
+package org.oppia.android.util.parser.html
+
+import javax.inject.Qualifier
+
+/** Qualifier for injecting the string resource ID of the worked example question label. */
+@Qualifier
+annotation class WorkedExampleQuestionLabelStringId
+
+/** Qualifier for injecting the string resource ID of the worked example answer label. */
+@Qualifier
+annotation class WorkedExampleAnswerLabelStringId
